Moved API into Separate Project
Also, managed to find a way to not have to manually copy over all csproj's in docker file, so creating / Renaming new ones going forward will be much easier
