+++ api/Dockerfile
@@ -1,20 +1,11 @@
 FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
 
 WORKDIR /src
-COPY ["ExpressedRealms.sln", "."]
-COPY ["ExpressedRealms.Repositories.Shared/ExpressedRealms.Repositories.Shared.csproj", "ExpressedRealms.Repositories.Shared/"]
-COPY ["ExpressedRealms.Email/ExpressedRealms.Email.csproj", "ExpressedRealms.Email/"]
-COPY ["ExpressedRealms.Repositories.Characters/ExpressedRealms.Repositories.Characters.csproj", "ExpressedRealms.Repositories.Characters/"]
-COPY ["ExpressedRealms.Repositories.Expressions/ExpressedRealms.Repositories.Expressions.csproj", "ExpressedRealms.Repositories.Expressions/"]
-COPY ["ExpressedRealms.Repositories.Admin/ExpressedRealms.Repositories.Admin.csproj", "ExpressedRealms.Repositories.Admin/"]
-COPY ["ExpressedRealms.Powers.Repository/ExpressedRealms.Powers.Repository.csproj", "ExpressedRealms.Powers.Repository/"]
-COPY ["ExpressedRealms.DB/ExpressedRealms.DB.csproj", "ExpressedRealms.DB/"]
-COPY ["ExpressedRealms.MigrationProject/ExpressedRealms.MigrationProject.csproj", "ExpressedRealms.MigrationProject/"]
-COPY ["ExpressedRealms.Authentication/ExpressedRealms.Authentication.csproj", "ExpressedRealms.Authentication/"]
-COPY ["ExpressedRealms.Server/ExpressedRealms.Server.csproj", "ExpressedRealms.Server/"]
-COPY ["ExpressedRealms.FeatureFlags/ExpressedRealms.FeatureFlags.csproj", "ExpressedRealms.FeatureFlags/"]
-COPY ["ExpressedRealms.Powers.API/ExpressedRealms.Powers.API.csproj", "ExpressedRealms.Powers.API/"]
-COPY ["ExpressedRealms.Server.Shared/ExpressedRealms.Server.Shared.csproj", "ExpressedRealms.Server.Shared/"]
+COPY ["ExpressedRealms.sln", "./"]
+# https://stackoverflow.com/a/51373386
+COPY */*.csproj ./
+RUN for file in $(ls *.csproj); do mkdir -p ${file%.*}/ && mv $file ${file%.*}/; done
+
 RUN dotnet restore
 
 COPY . .
