+++ api/ExpressedRealms.Characters.API/CharacterEndPoints/Responses/CharacterResponse.cs
@@ -1,4 +1,4 @@
-namespace ExpressedRealms.Server.EndPoints.CharacterEndPoints.DTOs;
+namespace ExpressedRealms.Characters.API.CharacterEndPoints.Responses;
 
 public class CharacterResponse
 {
