+++ api/ExpressedRealms.Characters.API/CharacterEndPoints/Responses/FactionOptionResponse.cs
@@ -1,3 +1,3 @@
-namespace ExpressedRealms.Server.EndPoints.CharacterEndPoints.Responses;
+namespace ExpressedRealms.Characters.API.CharacterEndPoints.Responses;
 
 public record FactionOptionResponse(int Id, string Name, string Description);
