+++ api/ExpressedRealms.Characters.API/CharacterEndPoints/StatDTOs/EditStatDTO.cs
@@ -1,6 +1,6 @@
 using ExpressedRealms.Characters.Repository.Stats.Enums;
 
-namespace ExpressedRealms.Server.EndPoints.CharacterEndPoints.StatDTOs;
+namespace ExpressedRealms.Characters.API.CharacterEndPoints.StatDTOs;
 
 public class EditStatDTO
 {
