+++ api/ExpressedRealms.Characters.API/CharacterEndPoints/StatDTOs/EditStatRequest.cs
@@ -1,5 +1,5 @@
 using ExpressedRealms.Characters.Repository.Stats.Enums;
 
-namespace ExpressedRealms.Server.EndPoints.CharacterEndPoints.StatDTOs;
+namespace ExpressedRealms.Characters.API.CharacterEndPoints.StatDTOs;
 
 public record EditStatRequest(int CharacterId, StatType StatTypeId);
