+++ api/ExpressedRealms.Characters.API/CharacterEndPoints/StatDTOs/SingleStatInfo.cs
@@ -1,6 +1,6 @@
 using ExpressedRealms.Characters.Repository.Stats.Enums;
 
-namespace ExpressedRealms.Server.EndPoints.CharacterEndPoints.StatDTOs;
+namespace ExpressedRealms.Characters.API.CharacterEndPoints.StatDTOs;
 
 public record SingleStatInfo
 {
