+++ api/ExpressedRealms.Characters.API/CharacterEndPoints/StatDTOs/SmallStatInfo.cs
@@ -1,6 +1,6 @@
 using ExpressedRealms.Characters.Repository.Stats.Enums;
 
-namespace ExpressedRealms.Server.EndPoints.CharacterEndPoints.StatDTOs;
+namespace ExpressedRealms.Characters.API.CharacterEndPoints.StatDTOs;
 
 public class SmallStatInfo(Characters.Repository.Stats.DTOs.SmallStatInfo smallStatInfo)
 {
