+++ api/ExpressedRealms.Server/EndPoints/NavigationEndpoints/NavigationEndpoints.cs
@@ -5,8 +5,8 @@ using ExpressedRealms.FeatureFlags.FeatureClient;
 using ExpressedRealms.Repositories.Expressions.Expressions;
 using ExpressedRealms.Server.EndPoints.NavigationEndpoints.DTOs;
 using ExpressedRealms.Server.EndPoints.NavigationEndpoints.Responses;
-using ExpressedRealms.Server.Extensions;
 using ExpressedRealms.Server.Shared;
+using ExpressedRealms.Server.Shared.Extensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.EntityFrameworkCore;
