+++ api/ExpressedRealms.Server/ExpressedRealms.Server.csproj
@@ -30,6 +30,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ExpressedRealms.Authentication\ExpressedRealms.Authentication.csproj" />
+    <ProjectReference Include="..\ExpressedRealms.Characters.API\ExpressedRealms.Characters.API.csproj" />
     <ProjectReference Include="..\ExpressedRealms.DB\ExpressedRealms.DB.csproj" />
     <ProjectReference Include="..\ExpressedRealms.Email\ExpressedRealms.Email.csproj" />
     <ProjectReference Include="..\ExpressedRealms.FeatureFlags\ExpressedRealms.FeatureFlags.csproj" />
