+++ api/ExpressedRealms.sln
@@ -29,6 +29,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExpressedRealms.Powers.API"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExpressedRealms.Server.Shared", "ExpressedRealms.Server.Shared\ExpressedRealms.Server.Shared.csproj", "{FE6C7E51-E450-4A5C-B58E-A08584BB9FC6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExpressedRealms.Characters.API", "ExpressedRealms.Characters.API\ExpressedRealms.Characters.API.csproj", "{22B275D6-19D4-4859-8159-2662DA34E6C0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -87,6 +89,10 @@ Global
 		{FE6C7E51-E450-4A5C-B58E-A08584BB9FC6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FE6C7E51-E450-4A5C-B58E-A08584BB9FC6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FE6C7E51-E450-4A5C-B58E-A08584BB9FC6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{22B275D6-19D4-4859-8159-2662DA34E6C0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{22B275D6-19D4-4859-8159-2662DA34E6C0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{22B275D6-19D4-4859-8159-2662DA34E6C0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{22B275D6-19D4-4859-8159-2662DA34E6C0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
